refactor: 디렉터리면 offline_pay에서 offline으로 변경
