+++ src/main/java/org/phoenix/planet/consumer/ReceiptEventConsumer.java
@@ -8,7 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.phoenix.planet.constant.KafkaTopic;
 import org.phoenix.planet.dto.eco_stock.raw.EcoStock;
-import org.phoenix.planet.dto.offline_pay.request.OfflinePayload;
+import org.phoenix.planet.dto.offline.request.OfflinePayload;
 import org.phoenix.planet.service.card.MemberCardService;
 import org.phoenix.planet.service.eco_stock.EcoStockIssueService;
 import org.phoenix.planet.service.eco_stock.EcoStockService;
