+++ src/main/java/org/phoenix/planet/controller/OfflinePayController.java
@@ -1,7 +1,7 @@
 package org.phoenix.planet.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.phoenix.planet.dto.offline_pay.request.OfflinePayload;
+import org.phoenix.planet.dto.offline.request.OfflinePayload;
 import org.phoenix.planet.service.receipt.OfflinePayService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
