+++ src/main/java/org/phoenix/planet/service/receipt/OfflinePayService.java
@@ -1,6 +1,6 @@
 package org.phoenix.planet.service.receipt;
 
-import org.phoenix.planet.dto.offline_pay.request.OfflinePayload;
+import org.phoenix.planet.dto.offline.request.OfflinePayload;
 
 public interface OfflinePayService {
